refactor: extract branch and access actions from AgentWorkspace
